+++ pose-detection/src/blazepose_mediapipe/mediapipe_test.ts
@@ -23,7 +23,7 @@ import {expectArraysClose} from '@tensorflow/tfjs-core/dist/test_util';
 import * as poseDetection from '../index';
 import {toImageDataLossy} from '../shared/calculators/mask_util';
 import {imageToBooleanMask, KARMA_SERVER, loadImage, segmentationIOU} from '../shared/test_util';
-import {EXPECTED_LANDMARKS, EXPECTED_WORLD_LANDMARKS, getXYPerFrame, loadVideo} from '../test_util';
+import {getXYPerFrame, loadVideo} from '../test_util';
 
 import {BlazePoseMediaPipeModelConfig} from './types';
 
@@ -43,6 +43,30 @@ const EPSILON_VIDEO = 31;
 const EPSILON_VIDEO_WORLD = 0.24;
 // Measured in percent.
 const EPSILON_IOU = 0.88;
+
+// ref:
+// https://github.com/google/mediapipe/blob/7c331ad58b2cca0dca468e342768900041d65adc/mediapipe/python/solutions/pose_test.py#L31-L51
+export const EXPECTED_LANDMARKS = [
+  [460, 283], [467, 273], [471, 273], [474, 273], [465, 273], [465, 273],
+  [466, 273], [491, 277], [480, 277], [470, 294], [465, 294], [545, 319],
+  [453, 329], [622, 323], [375, 316], [696, 316], [299, 307], [719, 316],
+  [278, 306], [721, 311], [274, 304], [713, 313], [283, 306], [520, 476],
+  [467, 471], [612, 550], [358, 490], [701, 613], [349, 611], [709, 624],
+  [363, 630], [730, 633], [303, 628]
+];
+export const EXPECTED_WORLD_LANDMARKS = [
+  [-0.11, -0.59, -0.15], [-0.09, -0.64, -0.16], [-0.09, -0.64, -0.16],
+  [-0.09, -0.64, -0.16], [-0.11, -0.64, -0.14], [-0.11, -0.64, -0.14],
+  [-0.11, -0.64, -0.14], [0.01, -0.65, -0.15],  [-0.06, -0.64, -0.05],
+  [-0.07, -0.57, -0.15], [-0.09, -0.57, -0.12], [0.18, -0.49, -0.09],
+  [-0.14, -0.5, -0.03],  [0.41, -0.48, -0.11],  [-0.42, -0.5, -0.02],
+  [0.64, -0.49, -0.17],  [-0.63, -0.51, -0.13], [0.7, -0.5, -0.19],
+  [-0.71, -0.53, -0.15], [0.72, -0.51, -0.23],  [-0.69, -0.54, -0.19],
+  [0.66, -0.49, -0.19],  [-0.64, -0.52, -0.15], [0.09, 0., -0.04],
+  [-0.09, -0., 0.03],    [0.41, 0.23, -0.09],   [-0.43, 0.1, -0.11],
+  [0.69, 0.49, -0.04],   [-0.48, 0.47, -0.02],  [0.72, 0.52, -0.04],
+  [-0.48, 0.51, -0.02],  [0.8, 0.5, -0.14],     [-0.59, 0.52, -0.11],
+];
 
 describeWithFlags('MediaPipe Pose static image ', BROWSER_ENVS, () => {
   let detector: poseDetection.PoseDetector;
